Evaluation questions
A set of 100 evaluation questions set as a benchmark to evaluate the models. The questions are divided into fiveA more detailed analysis of how each model's performance changes with the complexity of the task was possible, providing insights into their strengths and weaknesses.
o	Category 1: Simple Retrieval (25 questions) assesses fundamental selection and filtering techniques. An analyst would utilise these fundamental queries to swiftly extract a particular piece of data. Since only one value is expected, this category is an ideal way to test the models' most basic data-grounding skills. "What was the closing price on 2023-10-26?" is one example.
o	Category 2: The application of common aggregate functions, such as mean (), sum (), and max(), is tested in Category 2: Basic Aggregation (25 questions). The difficulty of these queries increases because they call for the model to do more than just receive data; it also needs to do a basic calculation on it. "What was the average closing price in 2021?" is one example.
o	Category 3: More intricate, multi-part Boolean logic is needed for Conditional Analysis (25 questions). The ability of a model to manage conditional filtering—a fundamental ability for any data processing task—is tested by these questions. For instance, "Retrieve all rows where High > 150."
o	Category 4: Multi-step processes, like generating new columns from preexisting ones or calculating moving averages using the rolling () function, are tested in the 15-question Complex Calculation test. A model's capacity for sophisticated thinking and the ability to combine several operations into a coherent response are put to the test in this area.
o	Category 5: Visualisation (10 questions) is Category 5. evaluates the ability to produce accurate matplotlib code for making basic plots, such as bar and line charts. This category is intended to assess a model's ability to generate code and be creative in addition to retrieving data.
 categories